Manual_Search_AI_assistant/requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
transformers
requests
